qnn/requirements.txt
absl-py==2.1.0
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.51.0
fsspec==2024.2.0
grpcio==1.63.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==24.0
pandas==2.2.2
pillow==10.2.0
protobuf==5.26.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.4.2
scipy==1.13.0
setuptools==69.5.1
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
torch==2.3.0+cpu
torchaudio==2.3.0+cpu
torchvision==0.18.0+cpu
tqdm==4.66.4
typing_extensions==4.9.0
tzdata==2024.1
Werkzeug==3.0.3
